+++ Questionable/Domain/QuestInfo.cs
@@ -108,7 +108,11 @@ internal sealed class QuestInfo : IQuestInfo
                 return item != null ? ItemReward.CreateFromItem(item.Value, QuestId) : null;
             }).OfType<ItemReward>().ToList();
         Expansion = (EExpansionVersion)quest.Expansion.RowId;
+        AvailableIcon = quest.EventIconType.ValueNullable?.MapIconAvailable + 1;
+        InvalidIcon = quest.EventIconType.ValueNullable?.MapIconInvalid + 1;
     }
+    public uint? AvailableIcon { get; private set; }
+    public uint? InvalidIcon { get; private set; }
     public ImmutableList<QQuestId> QuestLocks { get; private set; }
     public EQuestJoin QuestLockJoin { get; private set; }
     public List<string> ActionUnlock { get; }