+++ Questionable/Windows/JournalComponents/QuestJournalComponent.cs
@@ -216,7 +216,6 @@ internal sealed class QuestJournalComponent
 
         ImGui.TableNextColumn();
         float spacing;
-        // ReSharper disable once UnusedVariable
         using (IDisposable font = pluginInterface.UiBuilder.IconFontFixedWidthHandle.Push())
         {
             spacing = ImGui.GetColumnWidth() / 2 - ImGui.CalcTextSize(FontAwesomeIcon.Check.ToIconString()).X;
@@ -273,7 +272,22 @@ internal sealed class QuestJournalComponent
 
         ImGui.TableNextColumn();
         (Vector4 color, FontAwesomeIcon icon, string text) = uiUtils.GetQuestStyle(questInfo.QuestId);
-        if (uiUtils.ChecklistItem(text.Split(':')[0], color, icon))
+        // FA icon override with quest icon
+        const uint CompleteIcon = 71025;
+        const uint QuestionIcon = 71026;
+        uint? iconOverride = null;
+        if (configuration.General.QuestIcons)
+        {
+            if (icon is FontAwesomeIcon.Running)
+                iconOverride = questInfo.AvailableIcon;
+            if (icon is FontAwesomeIcon.Times)
+                iconOverride = questInfo.InvalidIcon;
+            if (icon is FontAwesomeIcon.Check)
+                iconOverride = CompleteIcon;
+            if (icon is FontAwesomeIcon.QuestionCircle)
+                iconOverride = QuestionIcon;
+        }
+        if (uiUtils.ChecklistItem(text.Split(':')[0], color, icon, iconOverride: iconOverride))
             ImGui.SetTooltip(text);
     }