+++ Questionable/Windows/UiUtils.cs
@@ -6,7 +6,11 @@ using Questionable.Model.Questing;
 using Questionable.Windows.Common.Ui;
 namespace Questionable.Windows;
 
-internal sealed class UiUtils(QuestFunctions questFunctions, IDalamudPluginInterface pluginInterface, QuestData questData)
+internal sealed class UiUtils(
+    QuestFunctions questFunctions,
+    IDalamudPluginInterface pluginInterface,
+    QuestData questData,
+    ITextureProvider textureProvider)
 {
     public (Vector4 Color, FontAwesomeIcon Icon, string Status) GetQuestStyle(ElementId elementId)
     {
@@ -62,14 +66,25 @@ internal sealed class UiUtils(QuestFunctions questFunctions, IDalamudPluginInter
         return (QstTheme.Danger, FontAwesomeIcon.Times);
     }
 
-    public bool ChecklistItem(string text, Vector4 color, FontAwesomeIcon icon, float extraPadding = 0)
+    public bool ChecklistItem(string text, Vector4 color, FontAwesomeIcon icon, float extraPadding = 0, uint? iconOverride = null)
     {
         if (extraPadding > 0)
             ImGui.SetCursorPosX(ImGui.GetCursorPosX() + extraPadding);
-
-        using (pluginInterface.UiBuilder.IconFontFixedWidthHandle.Push())
+        if (iconOverride != null &&
+            textureProvider.TryGetFromGameIcon(new(iconOverride.Value), out var tex) &&
+            tex.TryGetWrap(out var texture, out _))
         {
-            ImGui.TextColored(color, icon.ToIconString());
+            var drawSize = new Vector2(ImGui.GetTextLineHeightWithSpacing());
+            ImGui.Image(texture.Handle, drawSize);
+            ImGui.SameLine();
+            texture.Dispose();
+        }
+        else
+        {
+            using (pluginInterface.UiBuilder.IconFontFixedWidthHandle.Push())
+            {
+                ImGui.TextColored(color, icon.ToIconString());
+            }
         }
 
         bool hover = ImGui.IsItemHovered();